Comments and turned PSO entirely int-based on mass

+++ Assets/Scripts/Portion.cs
@@ -6,7 +6,7 @@ public struct Portion : IVerbose
     public readonly Food FoodType { get; }
 
     public int Mass { get; set; }
-    private readonly float Multiplier => Mass / Food.MASS;
+    private readonly float Multiplier => (float)Mass / Food.MASS;
 
 
     public Portion(Food food)
